project2.js
# Chapter 2.4 - Accepting Donations

Even if this part might not be vital for our bot, it is always useful to have 
something very simple which will accept all donations for us. To do this first
we will need to identify what a donation fundamentally is: a donation is when
the user sends us items and we do not send any items back in the trade. Knowing
this we can use our knowledge to implement this very simple concept. 

We can add a very simple line to our even listner:
```js
manager.on('newOffer', (offer) => {
	[...]
  if(offer.itemsToGive.length == 0){
		offer.accept((err, status) => {
			if (err) {
				console.log(err);
			} else {
				console.log(`Donation accepted. Status: ${status}.`)
			}
		});
	}
});
```

With these very simple lines we are first checking if the array which contains our
items (offer.itemsToGive) contains 0 items. We are using .length to do this. This 
property will equal 0 only if offer.itemsToGive is empty or contains no elements.
If we think about our previous definition of a donation then this will indeed mean
that if this statement is true then the trade offer is a donation.